# proto/openapi.yaml
openapi: 3.0.3
info:
  title: Banking Platform API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/transactions/transfer:
    post:
      operationId: transfer
      parameters:
        - name: Idempotency-Key
          in: header
          schema:
            type: string
      responses:
        "200":
          description: Transfer state or pre-debit risk-control challenge
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/WalletTransaction"
                  - $ref: "#/components/schemas/TransferRiskResponse"
...
  /api/admin/risk:
    get:
      operationId: adminListRiskEvaluations
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: level
          in: query
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH, CRITICAL]
        - name: action
          in: query
          schema:
            type: string
            enum: [ALLOW, WARN_USER, STEP_UP_AUTH, MANUAL_REVIEW, BLOCK]
      responses:
        "200":
          description: Paginated risk evaluation list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskEvaluationPage"
  /api/admin/risk/{id}:
    get:
      operationId: adminGetRiskEvaluation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Risk evaluation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskEvaluation"
...
  /api/admin/dlq:
    get:
      operationId: adminInspectDlq
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Wallet event DLQ messages
...
  /api/support/chat/sessions/{sessionId}:
    get:
      operationId: getSupportChatSession
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Support chat transcript
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportSessionDetail"
  /api/support/chat/sessions/{sessionId}/messages:
    post:
      operationId: sendSupportChatMessage
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Policy-validated assistant answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendSupportMessageResponse"
  /api/support/chat/sessions/{sessionId}/handoff:
    post:
      operationId: requestSupportHandoff
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Human support handoff case
  /api/admin/support/cases:
    get:
      operationId: adminListSupportCases
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
        - name: topic
          in: query
          schema:
            type: string
        - name: q
          in: query
          description: Search support cases by case, session, user, related transaction, assignee, status, or topic
          schema:
            type: string
      responses:
        "200":
          description: Paginated support cases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportCasePage"
  /api/admin/support/cases/{caseId}:
    get:
      operationId: adminGetSupportCase
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Support case detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportCaseDetail"
...
components:
  schemas:
...
    MoneyString:
      type: string
      pattern: "^-?[0-9]+(\\.[0-9]+)?$"
    RiskLevel:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
    RecommendedAction:
      type: string
      enum: [ALLOW, WARN_USER, STEP_UP_AUTH, MANUAL_REVIEW, BLOCK]
    RiskReason:
      type: object
      required: [code, weight, message, evidence]
      properties:
        code:
          type: string
        weight:
          type: integer
        message:
          type: string
        evidence:
          type: object
          additionalProperties:
            type: string
    RiskDecision:
      type: object
      required: [riskEvaluationId, riskScore, riskLevel, recommendedAction, reasons, modelVersion, policyVersion, evaluatedAt, traceId]
      properties:
        riskEvaluationId:
          type: string
          format: uuid
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        recommendedAction:
          $ref: "#/components/schemas/RecommendedAction"
        reasons:
          type: array
          items:
            $ref: "#/components/schemas/RiskReason"
        modelVersion:
          type: string
        policyVersion:
          type: string
        evaluatedAt:
          type: string
          format: date-time
        traceId:
          type: string
          format: uuid
    TransferRiskResponse:
      allOf:
        - $ref: "#/components/schemas/RiskDecision"
        - type: object
          required: [result, message]
          properties:
            result:
              type: string
              enum: [RISK_WARNING_REQUIRED, RISK_STEP_UP_REQUIRED, RISK_MANUAL_REVIEW_REQUIRED, RISK_BLOCKED]
            transactionId:
              type: string
              format: uuid
              nullable: true
            message:
              type: string
    WalletTransaction:
      type: object
      required: [id, senderId, receiverId, amount, currency, status, idempotencyKey, createdAt, updatedAt, debitApplied]
      properties:
        id:
          type: string
          format: uuid
        senderId:
          type: string
          format: uuid
        receiverId:
          type: string
          format: uuid
        amount:
          $ref: "#/components/schemas/MoneyString"
        currency:
          type: string
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED, COMPENSATING, CANCELLED]
        idempotencyKey:
          type: string
        correlationId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        debitApplied:
          type: boolean
        note:
          type: string
          nullable: true
        failureReason:
          type: string
          nullable: true
        reviewStatus:
          type: string
          nullable: true
        riskEvaluationId:
          type: string
          format: uuid
          nullable: true
    RiskEvaluation:
      type: object
      required: [id, senderAccountId, receiverAccountId, idempotencyKey, payloadHash, amount, currency, riskScore, riskLevel, recommendedAction, reasons, features, modelVersion, policyVersion, decisionStatus, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        transactionId:
          type: string
          format: uuid
          nullable: true
        senderAccountId:
          type: string
          format: uuid
        receiverAccountId:
          type: string
          format: uuid
          nullable: true
        idempotencyKey:
          type: string
        payloadHash:
          type: string
        amount:
          $ref: "#/components/schemas/MoneyString"
        currency:
          type: string
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        recommendedAction:
          $ref: "#/components/schemas/RecommendedAction"
        reasons:
          type: array
          items:
            $ref: "#/components/schemas/RiskReason"
        features:
          type: object
          additionalProperties:
            type: string
        modelVersion:
          type: string
        policyVersion:
          type: string
        decisionStatus:
          type: string
        traceId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RiskEvaluationPage:
      type: object
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RiskEvaluation"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
...
    SupportSuggestedAction:
      type: object
      required: [type, label]
      properties:
        type:
          type: string
          enum: [OPEN_TRANSACTION_DETAIL, OPEN_HISTORY, OPEN_PIN_SETTINGS, CONTACT_HUMAN_SUPPORT, REFRESH_TRANSACTION_STATUS, OPEN_SECURITY_SETTINGS]
        label:
          type: string
        targetId:
          type: string
          format: uuid
          nullable: true
...
    SendSupportMessageResponse:
      type: object
      required: [messageId, answer, citations, suggestedActions, traceId]
      properties:
        messageId:
          type: string
          format: uuid
        answer:
          type: string
        citations:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              id:
                type: string
                format: uuid
              fields:
                type: array
                items:
                  type: string
        suggestedActions:
          type: array
          items:
            $ref: "#/components/schemas/SupportSuggestedAction"
        traceId:
          type: string
          format: uuid
    SupportChatMessage:
      type: object
      required: [id, sessionId, senderType, message, createdAt]
      properties:
        id:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        senderType:
          type: string
          enum: [USER, AI, ADMIN, SYSTEM]
        message:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
        createdAt:
          type: string
          format: date-time
    SupportSessionDetail:
      type: object
      required: [sessionId, status, topic, messages]
      properties:
        sessionId:
          type: string
          format: uuid
        status:
          type: string
        topic:
          type: string
        relatedTransactionId:
          type: string
          format: uuid
          nullable: true
        messages:
          type: array
          items:
            $ref: "#/components/schemas/SupportChatMessage"
    SupportCase:
      type: object
      required: [caseId, sessionId, userId, status, topic, createdAt, updatedAt]
      properties:
        caseId:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
        topic:
          type: string
        relatedTransactionId:
          type: string
          format: uuid
          nullable: true
        assignedAdminId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SupportCasePage:
      type: object
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SupportCase"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    SupportTransactionSnapshot:
      type: object
      required: [id, status, compensated]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        failureReason:
          type: string
          nullable: true
        compensated:
          type: boolean
        traceId:
          type: string
          format: uuid
          nullable: true
    SupportCaseDetail:
      type: object
      required: [caseId, sessionId, status, topic, summary, messages]
      properties:
        caseId:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        status:
          type: string
        topic:
          type: string
        relatedTransactionId:
          type: string
          format: uuid
          nullable: true
        summary:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/SupportChatMessage"
        transactionSnapshot:
          $ref: "#/components/schemas/SupportTransactionSnapshot"